third_party/blink/web_tests/http/tests/security/xss-exception-expected.txt

CONSOLE MESSAGE: PASS: not executed in iframe context
CONSOLE MESSAGE: PASS: not executed in iframe context
CONSOLE MESSAGE: PASS: not executed in iframe context
Test passes if three "PASS" messages are logged to the console. 
